+++ lib/src/all_native.dart
@@ -14,3 +14,5 @@ export 'delegates/flex_delegate_native.dart';
 export 'util/byte_conversion_utils_native.dart';
 export 'util/model_checkpoint.dart';
 export 'custom_ops/transpose_conv_bias_native.dart';
+export 'interpreter_factory.dart';
+export 'interpreter_pool.dart';